Respond to telegram bot messages
