+++ backend/telegram/src/service/mod.rs
@@ -1,690 +1,41 @@
-/*
-   The telegram service polls the telegram /getUpdates api deserializes the messages (that is knows how to handle) and publishes them to a channel.
+use tokio::sync::broadcast::Sender;
 
-   We expect to implement logic like this:
+use crate::TelegramClient;
 
-   A new chat_id is seen
-       - Create a new recipient for the chat id (via published update)
-       - Send a welcome message to the chat (including the chat id for reference)
+use self::processor::poll_get_updates;
 
-   An existing chat_id is seen (check if we need to update the chat name in the recipient)
-       - No message is required
+mod processor;
+mod responder;
 
-   A direct message is seen
-       - Send the message with the chat id
-*/
-
-/*
-Example API Polling responses
-https://core.telegram.org/bots/api#getupdates
-
-{
-    "ok": true,
-    "result": [
-        {
-            "update_id": 794348048,
-            "message": {
-                "message_id": 30,
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "chat": {
-                    "id": -903279238,
-                    "title": "User1 & bot-name",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "date": 1691531796,
-                "group_chat_created": true
-            }
-        },
-        {
-            "update_id": 794348049,
-            "my_chat_member": {
-                "chat": {
-                    "id": -903279238,
-                    "title": "User1 & bot-name",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "date": 1691531796,
-                "old_chat_member": {
-                    "user": {
-                        "id": 6544022299,
-                        "is_bot": true,
-                        "first_name": "bot-name",
-                        "username": "bot-name"
-                    },
-                    "status": "left"
-                },
-                "new_chat_member": {
-                    "user": {
-                        "id": 6544022299,
-                        "is_bot": true,
-                        "first_name": "bot-name",
-                        "username": "bot-name"
-                    },
-                    "status": "member"
-                }
-            }
-        },
-        {
-            "update_id": 794348050,
-            "message": {
-                "message_id": 31,
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "chat": {
-                    "id": -903279238,
-                    "title": "User1 & bot-name (West)",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "date": 1691531864,
-                "new_chat_title": "User1 & bot-name (West)"
-            }
-        },
-        {
-            "update_id": 794348051,
-            "message": {
-                "message_id": 32,
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "chat": {
-                    "id": -914917543,
-                    "title": "User1 & bot-name",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "date": 1691536017,
-                "text": "@bot-name Can you tell me this chat_id please?",
-                "entities": [
-                    {
-                        "offset": 0,
-                        "length": 15,
-                        "type": "mention"
-                    }
-                ]
-            }
-        },
-        {
-            "update_id": 794348052,
-            "message": {
-                "message_id": 33,
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "chat": {
-                    "id": -914917543,
-                    "title": "User1 & bot-name",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "date": 1691536034,
-                "text": "This is a normal non direct message message..."
-            }
-        },
-        {
-            "update_id": 794348053,
-            "message": {
-                "message_id": 34,
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "chat": {
-                    "id": -931768832,
-                    "title": "Testing Notifications :)",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "date": 1691536201,
-                "text": "User2 Just wanting to check how mentions look that aren\u2019t too the bot\u2026",
-                "entities": [
-                    {
-                        "offset": 0,
-                        "length": 5,
-                        "type": "text_mention",
-                        "user": {
-                            "id": 1320002934,
-                            "is_bot": false,
-                            "first_name": "User2",
-                            "last_name": "Last2"
-                        }
-                    }
-                ]
-            }
-        }
-    ]
-}
-*/
-
-use crate::{TelegramClient, TelegramUpdate, TelegramUpdateOrId};
-use log::{self};
-
-static TELEGRAM_POLL_TIMEOUT_SECONDS: i64 = 30;
-static ERROR_SLEEP_SECONDS: u64 = 10;
-
-pub async fn poll_get_updates(
-    telegram_client: &TelegramClient,
-    tx_updates: &tokio::sync::mpsc::Sender<TelegramUpdate>,
-) {
-    // BTW : Might be a good idea to persist last_update_id to the KV store so we can pickup from where we left off on a restart?
-    let mut last_update_id = None;
-
-    loop {
-        let updates = telegram_client
-            .get_updates(last_update_id, TELEGRAM_POLL_TIMEOUT_SECONDS)
-            .await;
-        match updates {
-            Ok(updates) => {
-                let num = updates.len();
-                log::debug!("Got {} updates", num);
-                if num > 0 {
-                    last_update_id = handle_json_updates(updates, tx_updates).await;
-                }
-            }
-            Err(error) => {
-                log::error!("Error getting updates: {:?} \n Sleeping...", error);
-                // Sleep for a bit so we don't hammer the CPU or telegram API
-                tokio::time::sleep(std::time::Duration::from_secs(ERROR_SLEEP_SECONDS)).await;
-            }
-        };
-    }
+pub struct TelegramService {
+    pub client: TelegramClient,
+    pub app_url: String,
+    pub updates_channel: Sender<crate::TelegramUpdate>,
 }
 
-async fn handle_json_updates(
-    updates: Vec<serde_json::Value>,
-    tx_updates: &tokio::sync::mpsc::Sender<TelegramUpdate>,
-) -> Option<i64> {
-    let mut last_update_id = None;
-    for update in updates {
-        // Now try to parse the update using serde_json
-        let telegram_update: TelegramUpdateOrId = match serde_json::from_value(update.clone()) {
-            Ok(telegram_update) => telegram_update,
-            Err(error) => {
-                log::error!("Error parsing update: {:?} update: {:?}", error, update);
-                continue;
-            }
-        };
-
-        let update_id = match telegram_update {
-            TelegramUpdateOrId::Update(telegram_update) => {
-                let update_id = telegram_update.update_id;
-                // Send the update on the channel so other processors can handle it.
-                let result = tx_updates.send(telegram_update).await;
-                match result {
-                    Ok(_) => {
-                        log::debug!("Sent update to tx_updates");
-                    }
-                    Err(error) => {
-                        log::error!("Error sending message to tx_updates: {:?}", error);
-                    }
-                };
-                update_id
-            }
-            TelegramUpdateOrId::Id(id) => id.update_id,
-        };
-
-        // Update the last_update_id if it is bigger that the one we just saw
-        last_update_id = match last_update_id {
-            Some(id) if id > update_id => Some(id),
-            _ => Some(update_id),
-        };
-    }
-    last_update_id
-}
-
-#[cfg(test)]
-mod test {
-    use crate::{service::handle_json_updates, TelegramUpdate};
-
-    #[tokio::test]
-    async fn test_handle_json_updates_empty_vec() {
-        let json = r#"[]"#;
-        let updates: Vec<serde_json::Value> = serde_json::from_str(json).unwrap();
-        const TELEGRAM_UPDATE_BUFFER_SIZE: usize = 8;
-        let (tx, _rx) = tokio::sync::mpsc::channel::<TelegramUpdate>(TELEGRAM_UPDATE_BUFFER_SIZE);
-
-        let last_update_id = handle_json_updates(updates, &tx).await;
-
-        assert!(last_update_id == None);
-    }
-
-    #[tokio::test]
-    async fn test_handle_json_updates_empty_message() {
-        let json = r#"        [
-         {
-            "update_id": 1111,
-            "something_else": {
-                "date": 1691536034,
-                "text": "We don't know what this is, but we do have a update_id"
-            }
+impl TelegramService {
+    pub fn new(client: TelegramClient, app_url: String) -> Self {
+        let (tx, _) = tokio::sync::broadcast::channel(10);
+        TelegramService {
+            client,
+            app_url,
+            updates_channel: tx, // we only need the transmission side because consumers just subscribe via this
         }
-        ]"#;
-        let updates: Vec<serde_json::Value> = serde_json::from_str(json).unwrap();
-        const TELEGRAM_UPDATE_BUFFER_SIZE: usize = 8;
-        let (tx, _rx) = tokio::sync::mpsc::channel::<TelegramUpdate>(TELEGRAM_UPDATE_BUFFER_SIZE);
-
-        let last_update_id = handle_json_updates(updates, &tx).await;
-
-        assert!(last_update_id == Some(1111));
     }
 
-    #[tokio::test]
-    async fn test_handle_json_updates_invalid_message() {
-        let json = r#"        [
-         {
-            "update_id": 1111,
-             "message": {
-                "message_id": 4444,
-                "date": 1691536034,
-                "text": "This is a direct message message..."
-            }
-        }
-        ]"#;
-        let updates: Vec<serde_json::Value> = serde_json::from_str(json).unwrap();
-        const TELEGRAM_UPDATE_BUFFER_SIZE: usize = 8;
-        let (tx, _rx) = tokio::sync::mpsc::channel::<TelegramUpdate>(TELEGRAM_UPDATE_BUFFER_SIZE);
+    pub async fn init(self) -> Sender<crate::TelegramUpdate> {
+        // let app_url = self.app_url.clone();
+        let responder_channel = self.updates_channel.clone();
+        let responder_client = self.client.clone();
+        tokio::spawn(async move {
+            responder::handle_telegram_updates(&responder_client, &responder_channel).await;
+        });
 
-        let last_update_id = handle_json_updates(updates, &tx).await;
-
-        assert_eq!(last_update_id, Some(1111));
-    }
-
-    #[tokio::test]
-    async fn test_handle_json_updates_single_private_message() {
-        let json = r#"
-        [
-         {
-            "update_id": 1111,
-            "message": {
-                "message_id": 4444,
-                "from": {
-                    "id": 5555,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "chat": {
-                    "id": 5068627745,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "type": "private"
-                },
-                "date": 1691536034,
-                "text": "This is a direct message message..."
-            }
-        }
-        ]"#;
-        let updates: Vec<serde_json::Value> = serde_json::from_str(json).unwrap();
-        const TELEGRAM_UPDATE_BUFFER_SIZE: usize = 8;
-        let (tx, mut rx) =
-            tokio::sync::mpsc::channel::<TelegramUpdate>(TELEGRAM_UPDATE_BUFFER_SIZE);
-
-        let last_update_id = handle_json_updates(updates, &tx).await;
-
-        assert!(last_update_id == Some(1111));
-
-        // Get the message from the channel
-        let telegram_update = rx.recv().await.unwrap();
-        assert!(telegram_update.message.is_some());
-        let message = telegram_update.clone().message.unwrap();
-        assert_eq!(message.message_id, 4444);
-        assert_eq!(message.from.id, 5555);
-        assert!(message.from.is_bot == false);
-        assert_eq!(
-            telegram_update.chat().unwrap().name(),
-            "Telegram: User1 Last1".to_string()
-        );
-        assert_eq!(message.text.unwrap(), "This is a direct message message...");
-    }
-
-    #[tokio::test]
-    async fn test_handle_json_updates_single_message() {
-        let json = r#"
-        [
-         {
-            "update_id": 794348052,
-            "message": {
-                "message_id": 33,
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "chat": {
-                    "id": -914917543,
-                    "title": "User1 & bot-name",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "date": 1691536034,
-                "text": "This is a normal non direct message message..."
-            }
-        }
-        ]"#;
-        let updates: Vec<serde_json::Value> = serde_json::from_str(json).unwrap();
-        const TELEGRAM_UPDATE_BUFFER_SIZE: usize = 8;
-        let (tx, mut rx) =
-            tokio::sync::mpsc::channel::<TelegramUpdate>(TELEGRAM_UPDATE_BUFFER_SIZE);
-
-        let last_update_id = handle_json_updates(updates, &tx).await;
-
-        assert!(last_update_id == Some(794348052));
-
-        // Get the message from the channel
-        let telegram_update = rx.recv().await.unwrap();
-        assert!(telegram_update.message.is_some());
-        let message = telegram_update.clone().message.unwrap();
-        assert_eq!(message.message_id, 33);
-        assert_eq!(message.from.id, 5068627745);
-        assert!(message.from.is_bot == false);
-        assert_eq!(
-            telegram_update.chat().unwrap().name(),
-            "User1 & bot-name".to_string()
-        );
-        assert_eq!(
-            message.text.unwrap(),
-            "This is a normal non direct message message..."
-        );
-    }
-
-    #[tokio::test]
-    async fn test_handle_json_updates_single_my_chat_member_update() {
-        let json = r#"
-        [
-        {
-            "update_id": 794348049,
-            "my_chat_member": {
-                "chat": {
-                    "id": -903279238,
-                    "title": "User1 & bot-name",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "date": 1691531796,
-                "old_chat_member": {
-                    "user": {
-                        "id": 6544022299,
-                        "is_bot": true,
-                        "first_name": "bot-name",
-                        "username": "bot-name"
-                    },
-                    "status": "left"
-                },
-                "new_chat_member": {
-                    "user": {
-                        "id": 6544022299,
-                        "is_bot": true,
-                        "first_name": "bot-name",
-                        "username": "bot-name"
-                    },
-                    "status": "member"
-                }
-            }
-        }
-        ]"#;
-        let updates: Vec<serde_json::Value> = serde_json::from_str(json).unwrap();
-        const TELEGRAM_UPDATE_BUFFER_SIZE: usize = 8;
-        let (tx, _rx) = tokio::sync::mpsc::channel::<TelegramUpdate>(TELEGRAM_UPDATE_BUFFER_SIZE);
-
-        let last_update_id = handle_json_updates(updates, &tx).await;
-
-        assert!(last_update_id == Some(794348049));
-    }
-
-    #[tokio::test]
-    async fn test_handle_json_updates_lots_of_messages() {
-        let json = r#"
-       [
-        {
-            "update_id": 794348048,
-            "message": {
-                "message_id": 30,
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "chat": {
-                    "id": -903279238,
-                    "title": "User1 & bot-name",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "date": 1691531796,
-                "group_chat_created": true
-            }
-        },
-        {
-            "update_id": 794348049,
-            "my_chat_member": {
-                "chat": {
-                    "id": -903279238,
-                    "title": "User1 & bot-name",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "date": 1691531796,
-                "old_chat_member": {
-                    "user": {
-                        "id": 6544022299,
-                        "is_bot": true,
-                        "first_name": "bot-name",
-                        "username": "bot-name"
-                    },
-                    "status": "left"
-                },
-                "new_chat_member": {
-                    "user": {
-                        "id": 6544022299,
-                        "is_bot": true,
-                        "first_name": "bot-name",
-                        "username": "bot-name"
-                    },
-                    "status": "member"
-                }
-            }
-        },
-        {
-            "update_id": 794348050,
-            "message": {
-                "message_id": 31,
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "chat": {
-                    "id": -903279238,
-                    "title": "User1 & bot-name (West)",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "date": 1691531864,
-                "new_chat_title": "User1 & bot-name (West)"
-            }
-        },
-        {
-            "update_id": 794348051,
-            "message": {
-                "message_id": 32,
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "chat": {
-                    "id": -914917543,
-                    "title": "User1 & bot-name",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "date": 1691536017,
-                "text": "@bot-name Can you tell me this chat_id please?",
-                "entities": [
-                    {
-                        "offset": 0,
-                        "length": 15,
-                        "type": "mention"
-                    }
-                ]
-            }
-        },
-        {
-            "update_id": 794348052,
-            "message": {
-                "message_id": 33,
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "chat": {
-                    "id": -914917543,
-                    "title": "User1 & bot-name",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "date": 1691536034,
-                "text": "This is a normal non direct message message..."
-            }
-        },
-        {
-            "update_id": 794348053,
-            "message": {
-                "message_id": 34,
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "chat": {
-                    "id": -931768832,
-                    "title": "Testing Notifications :)",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "date": 1691536201,
-                "text": "User2 Just wanting to check how mentions look that aren\u2019t too the bot\u2026",
-                "entities": [
-                    {
-                        "offset": 0,
-                        "length": 5,
-                        "type": "text_mention",
-                        "user": {
-                            "id": 1320002934,
-                            "is_bot": false,
-                            "first_name": "User2",
-                            "last_name": "Last2"
-                        }
-                    }
-                ]
-            }
-        }
-    ]"#;
-        let updates: Vec<serde_json::Value> = serde_json::from_str(json).unwrap();
-        const TELEGRAM_UPDATE_BUFFER_SIZE: usize = 8;
-        let (tx, _rx) = tokio::sync::mpsc::channel::<TelegramUpdate>(TELEGRAM_UPDATE_BUFFER_SIZE);
-
-        let last_update_id = handle_json_updates(updates, &tx).await;
-
-        assert!(last_update_id == Some(794348053));
-    }
-
-    #[tokio::test]
-    async fn test_handle_json_updates_bad_data() {
-        let json = r#"[{
-            "update_id": 794348052,
-            "message": {
-                "message_id": 33,
-                "from": {
-                    "id": 5068627745,
-                    "is_bot": false,
-                    "first_name": "User1",
-                    "last_name": "Last1",
-                    "language_code": "en"
-                },
-                "chat": {
-                    "id": -914917543,
-                    "title": "User1 & bot-name",
-                    "type": "group",
-                    "all_members_are_administrators": true
-                },
-                "date": 1691536034,
-                "text": "This is a normal non direct message message..."
-            }
-            },{
-            "update_id": "Not a number",
-            "not_a_message": {
-                "not_a_message": 33,
-                "date": 1691536034,
-                "text": "This is a not a message!"
-            }
-        }]"#;
-        let updates: Vec<serde_json::Value> = serde_json::from_str(json).unwrap();
-        const TELEGRAM_UPDATE_BUFFER_SIZE: usize = 8;
-        let (tx, _rx) = tokio::sync::mpsc::channel::<TelegramUpdate>(TELEGRAM_UPDATE_BUFFER_SIZE);
-
-        let last_update_id = handle_json_updates(updates, &tx).await;
-
-        // We quietly skip the bad update_id
-        assert_eq!(last_update_id, Some(794348052));
+        let sender_channel = self.updates_channel.clone();
+        tokio::spawn(async move {
+            poll_get_updates(&self.client, &sender_channel).await;
+        });
+        // Return the channel so consumers can subscribe
+        self.updates_channel.clone()
     }
 }
